# Reduce variants to used variables before expanding (~300x faster)