This fixes https://github.com/conda-forge/conda-smithy/issues/2235

`get_package_combined_spec` computes the cartesian product of all variant lists, whether they are used or not. `render_recipe` then filters this list back down to the used values, immediately eliminating all of the unused variants.

By excluding unused variables before expanding the matrix, this eliminates the impact of unused variable dimensions on computing the variant result. The unused variables are put back into the used variant dicts, to ensure that the final return value is unchanged.

For example, this reduces the `conda-smithy rerender` time of petsc-feedstock from over 5 minutes to under 5 seconds (with no change in output), because `rattler_get_package_variants` is called to compute over 100k variants to produce 44 outputs. After this PR, it only computes a total of ~7k, a reduction of over 90%. petsc4py adds `python` to this same variant matrix, and also rerenders in about 6 seconds. I think it takes around 30 minutes to rerender without this PR, but I didn't want to wait for it to finish to get final numbers.

You can check [this script](https://gist.github.com/minrk/ce3dcb20be29efdc72da7184197ccd56) to render a simple recipe with increasing unused dimensions (as seen in conda-smithy) to see the difference:

<details>
<summary>before</summary>

```
dimension: 3
render_recipe produced 10 variants in 0.2s
render produced 10 variants in 0.1s
render computed a total of 200 package_variants
dimension: 4
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 400 package_variants
dimension: 5
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.2s
render computed a total of 800 package_variants
dimension: 6
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.2s
render computed a total of 1600 package_variants
dimension: 7
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.2s
render computed a total of 3200 package_variants
dimension: 8
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.4s
render computed a total of 6400 package_variants
dimension: 9
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 1.0s
render computed a total of 12800 package_variants
dimension: 10
render_recipe produced 10 variants in 0.2s
render produced 10 variants in 3.2s
render computed a total of 25600 package_variants
dimension: 11
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 11.8s
render computed a total of 51200 package_variants
dimension: 12
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 46.3s
render computed a total of 102400 package_variants
```

</details>

<details>
<summary>after</summary>

```
dimension: 3
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
dimension: 4
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
dimension: 5
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
dimension: 6
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
dimension: 7
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
dimension: 8
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
dimension: 9
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
dimension: 10
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
dimension: 11
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
dimension: 12
render_recipe produced 10 variants in 0.1s
render produced 10 variants in 0.1s
render computed a total of 100 package_variants
```

</details>