@minrk 2025-03-07
As a further optimization, it is not clear to me if [this per-metadata call to expand the variants](https://github.com/prefix-dev/rattler-build-conda-compat/pull/75/files#diff-77478d772015c6eb0acf745f154a3170ff44a128dca1e00997c90739d41c54dbR300) can produce different results, or if it could be called only once instead of for each of the variants, but I feel a lot less pressure to optimize a 6 second operation than a 38 minute one.